requirements.txt
Django>=3.2,<4.0
psycopg2-binary>=2.9,<3.0
djangorestframework
django-ckeditor
redis
googletrans==4.0.0-rc1
